pyproject.toml
[tool.poetry]
name = "transaction analysis"
version = "0.1.0"
description = ""
authors = ["Антонюк Евгений <38914583+AntonyukEvgeniy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pandas-stubs = "^2.2.3.241126"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"


[tool.poetry.group.dev.dependencies]
openpyxl = "^3.1.5"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
requests = "^2.32.3"
pandas = "^2.2.3"
pandas-stubs = "^2.2.3.241126"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.flake8]
# Check that this is aligned with your other tools like Black
max-line-length = 119
exclude = [
    # No need to traverse our git directory
    ".git",
    # There's no value in checking cache directories
    "__pycache__",
    ".venv"  # Добавляем .venv
]
# Use extend-ignore to add to already ignored checks which are anti-patterns like W503.
extend-ignore = [
    # PEP 8 recommends to treat : in slices as a binary operator with the lowest priority, and to leave an equal
    # amount of space on either side, except if a parameter is omitted (e.g. ham[1 + 1 :]).
    # This behaviour may raise E203 whitespace before ':' warnings in style guide enforcement tools like Flake8.
    # Since E203 is not PEP 8 compliant, we tell Flake8 to ignore this warning.
    # https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices
    "E203"
]

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'